Autodetect k8s compliance variant (#997)
* Autodetect k8s compliance variant

Add the ability to run k8s compliance checks either in
master (coresident with the k8s api server) or node (on any other node)
mode. When creating arguments for the module, use ps to look for any of
the process names for the api server. If found, run in master
mode. Otherwise, use ps to look for the process names for kubelet, which
indicate running on the non-master.

Also use this mechanism to silently not run a module at all, as compared
to running with an error. A new golang module interface ShouldRun
returns true if the module should run at all. For kube-bench, it runs if
either master or node programs could be found. For docker, it runs if
docker ps -q runs successfully.

* Just invoke ps directly.

No need to run it via a shell.

* Add ability to force a kube-bench variant

For testing purposes, it may be useful for force the kube-bench
compliance module to run in master or node mode instead of relying on
autodetect.

A new config option "security:
{compliance_kube_bench_variant: ("master"|"node")}" does this. If set,
the variant is passed as a param in the calendar, and the kube-bench
variant detection uses any provided variant instead of autodetecting.