Add auid unit tests (#998)
* Add unit tests for loginuid/name features

Add unit tests that test the functionality in
https://github.com/draios/sysdig/pull/1189, that adds the
loginuid/loginname as a filtercheck field.

Tests involve fetching the uid/loginuid independent from the events and
verifying they match up, both through direct syscalls as well as
indirectly through an exec.

* If capture fails, also print why.

This was useful debugging some unrelated test failures.

* Add an auid test for an explicit sudo

Add an auid test for an explicit sudo to the user nobody. Depends on
that user existing, and the test will skip if the user does not exist.