Change result json format (#999)
* Update comp_result message to match swagger api

Change the comp_result message such that most of the information is held
in the stringified json ext_result. This makes most of the other fields
redundant. We can't explicitly flag them as deprecated as we're using
protobuf 2, but add comments making them deprecated.

Also add a id field to the comp_task message so it can be included in
any later results.

* Add structs for canonical result formats

Add new structs TaskResultTest, TaskResultSection, ExtendedTaskResult
that generally map to the json expected in the swagger api. The only
difference is that these structs don't have descriptions.

Also update compclient to pass an id when starting artificial tasks.

* Update kube/docker scrpr to use new result structs

Update kube/docker scrapers to use the new result structs in module.go.

AssignRisk now works with the enum values instead of strings.

The other struct conversion is generally straightfoward, removing any
module-specific structs and replacing them with
ExtendedTaskResult/TaskResultSection/TaskResultTest.

For docker, now that there's a warn count/section, convert any warn
results to be considered warnings and not failures, and add statsd
metrics for the warn counts.

* Update test module/unit test to new result structs

Update the compliance unit test to use the new result structs. Generally
to verify results the ext_result needs to be json parsed to get the iter
counter, and an id needed to be added to all task definitions.

* Add Module name to Compliance Result

That way it won't be necessary to go through the calendar to determine
which module created a given result.

* Fix compile errors from conflict merge.