Minimal swarm support (#1010)
* Add minimal docker swarm support for policies.

Previously, we didn't support docker swarm metadata when scoping
security policies at the agent. Luckily, a lot of swarm metadata is
available as labels on the container, and label information is already
fetched.

For labels beginning with com.docker.swarm, drop the com.docker, and
attach the rest to the container group as a tag.

With this change, I was able to scope a policy to
swarm.service.name=helloworld, perform an action like "touch /bin/hack"
in a container that was a part of a swarm service named "helloworld and
had a docker label com.docker.swarm.service.name=helloworld, and the
policy event was triggered.

* Unit test on docker swarm scoped security policies

New unit test creates a single service that touches a sensitive file
repeatedly.

A new policy is linked to the falco rule "read_sensitive_file" and is
scoped to the swarm service name. The test looks for a policy event for
the scoped policy.