Use consistent environment for all commands (#1009)
When running modules, a custom environment is set up that changes things
like the docker socket location. However, this environment wasn't being
used when doing things like determining if a module should run.

In turn, this caused problems when doing the docker-bench ShouldRun
inside a container, as it was pointing at the wrong docker socket
location.

Change this by moving the creation of the custom environment into its
own method module.Env(), having the custom environment be set in the
ScheduledTask, and passing around the scheduled task in the interfaces
ShouldRun/GenArgs.