Include both details and items for compliance test (#1015)
Previously, if a compliance test had a list of items, the list was
included but the details was not. That meant you had a list of, say,
images, but no descriptive text of what the images were e.g. "Images w/o
HEALTHCHECK".

Change this so if items is non-empty, the text of details up to the
first colon is kept and included as the details. The frontend will
render both.