Fixed the libvirt test (#1023)
The systemd branch of the libvirt name checker does not seem to be
compatible with the naming scheme used by libvirt on my system. I
switched the equality operator with a find command, as the string it's
looking for is still in the name, it's just not the entire name.

It appears (looking at the algorithm that generates the sinsp name) as
though the generated name may be intentional rather than a bug, but it
might not expect the exact naming scheme my system comes up with. This
seems to me to be the cleanest way forward.