Only try to remove files in /opt/draios/lib when the directory exists
This is an edge case of the package split done in SMAGENT-1071.

When draios-agent is installed and then removed, /opt/draios/etc
and /opt/draios/run may stay behind. Then in the draios-agent-kmodule
script we check for existence of /opt/draios and then try to remove
files from /opt/draios/lib, which fails.

A workaround is to create /opt/draios/lib before removing the
draios-agent-kmodule package (or remove /opt/draios altogether).