Report failed connect() calls with unbound socket. SMAGENT-1105
When we detect a failing connection from a synchronous connect()
call, the source IP may not be available (still 0.0.0.0).

Adapt the logic in `sinsp_analyzer::emit_aggregated_connections()`
to handle this case.

Note: we still need to drop connections from 0.0.0.0 to 0.0.0.0
(it looks like they may be IPv6 connections from a quick check).