Use new grpc conns/clients for every action (#1030)
Previously, the security mgr used a single long-lived grpc connection
and grpc client for each of the main load, start, stop actions. We found
that after running an agent for a few days, the connection got in a bad
state where sending a new start rpc resulted in an immediate failure
from AsyncNext() that couldn't be recovered from.

To mitigate this, use a separate grpc connection for each
load/start/stop rpc as well as a new client.