Internal metrics statsd (#1033)
* Stop logging security_mgr metrics

Now that we're going to report security_mgr metrics, we need to clear
them every time they are sent, which means they can't be logged every 60
seconds. This wasn't that essential anyway, so simply remove it.

* Add ability to send some (not all) int metrics

Although the internal_metrics class collects lots of metrics, the
backend doesn't actually process them as they are in a part of the
metrics message not read by the backend.

Move them to the protos statsd_info section so they will be read by the
backend. As a part of this change, we don't want to actually send all
the metrics for now, as it may be too many for the backend to
handle. Allow this by adding a send_some() interface that implies
sending a subset of all the collected metrics.

This is also supported by external sources by adding another virtual
void send_some().

* Add ability to set tags for metrics

An alternate form of write_metric also passes a map of key/value
pairs. Those pairs are added to the metrics as tags.

* Collect policy-lvl metrics and report w/ send_some

Start collecting policy-level metrics such as policy event counts by
policy type and policy name (top 10 only), as well as an overall policy
count, and report them using send_some().

We also report metrics by policy severity. This is blocked on
https://sysdig.atlassian.net/browse/SSPROD-1147, so the implementation
is flexible, only setting a count by severity if the security policy
actually has a severity.

* Also add other useful agent metrics

We also want to report the sampling ratio, analyzer flush time, and
cointerface memory usage, so include them in the send_some() set.

* Const improvements.

Make various arguments const.