Add configurable descriptions to comp results (#1064)
New config option security: include_desc_compliance_results (default
true) controls whether or not to include descriptions in compliance
results. This is passed to the module mgr via the start message.

When true, the test section/test object includes the description.

Also include a result schema which helps identify the mapping from ids
to descriptions. This is related to
https://github.com/draios/api-swagger/pull/12, which updates the
definition of a ComplianceResult object to add the schema as a property.