Install jinja2 without virtualenv during build
When the `jinja_venv` target was executed before `sdchecks-deps`,
the resulting Python installation was broken, due to `virtualenv`
overwriting one of the files shipped with Python (the root cause
is unknown but might be related to the fact that we run Python out
of a nonstandard prefix and somehow virtualenv gets confused).

Fixed/worked around by avoiding the virtualenv completely
and installing jinja2 (for Dockerfile generation) directly into
the system Python path.