Autogenerate Python protobuf files on every protoanalyzer build
Rebuilding the draios_pb2.py file manually comes with a large risk
of forgetting to do it, so compile the protobufs directly in the
protoanalyzer image.

Note: I tried and failed to make it a multistage build, so that
protoc does not need to be included in the final image.