SMAGENT-1168: Re-enable variadic logging API (#1096)
* SMAGENT-1168: Re-enable variadic logging API

As part of the fix for SMAGENT-1163 , we backed out the variadic
logging API because because it caused problems when the length of
the log exceeded the fixed buffer size. This change re-enables
that API with support for log strings of arbitrary length.

Additionally, this change introduces a debug-only version of the
logging API.  Developers can the DBG_* version of the API to avoid
using macro wrappers around the normal API.

Finally, this change enables the logging API to be backed by either
the Poco logging facility or by a std::ostream.  This enables us to
write tests that exercise the logging API directly with the
expectation that the logs will be written to a stream that we can
later examine in the test.

* Add check for std::string w/ placholders

Added a UT test case where the logging API is invoked with a std::string
object that contains printf-style placeholders.  Verified that the
placeholders literals are printed (i.e., that nothing along the way
tried to fill those placeholders).

* Address Bryan's review comments