Environment variable blacklisting. SMAGENT-1170
If an environment variable matches any from a list of regexes,
exclude it from both the hashing and reporting to the backend.

Note: the list of regexes is not a vector because Poco::RegularExpression
is not copy-constructible (and not noexcept move-constructible)
so it cannot be placed in a vector. Holding `unique_ptr`s instead
of straight objects should work but in both cases we pay the cost
of dereferencing a pointer while iterating so this particular
choice shouldn't matter much.