SMAGENT-1175: Add -Wextra -Werror to agent build (#1098)
* SMAGENT-1175: Add -Wextra -Werror to agent build

Currently our builds include the -Wall compiler flag, which enables
some, but not all, warnings.  This change adds the -Wextra flag, which
enables some additional warnings.

This change also add -Werror, which will cause the compiler to report
any warning as an error (so now anything that generates a warning will
break the build).  This change explicitly omits some pervasive warnings
(e.g., unused parameters).

This change will require a parallel change in the sysdig component.

* Conditionally enable -Werror for sysdig

Update bootstrap-sysdig so that when we use it to build sysdig, we build
it with the option that enables -Werror.

* Make add the option to disable -Werror via env