Send result failed tasks (#1094)
* Add a notion of successful to comp_result

Now a result can be successful (or not). When successful=false,
failure_details will contain the full log of the failed task. Otherwise,
ext_result will continue to have the stringified json.

* Add an "always fail" module

Add a module for testing purposes that always fails due to the program
not-runnable not being executable. Used for unit tests.

* Track unsuccessful tasks and send as results

Sending failed results is configurable via security:
{compliance_send_failed_results: (true|false)}. Also make the current
default false, so if an agent is released before the backend/frontend
are ready it won't interfere. Once the backend/frontend are ready, we'll
change the default to true.

Make sure error messages are meaningful for all the possible reasons a
task can fail. This includes some cleanup of command logging to make the
output more human-readable.

Also, instead of writing command output to pipes and setting up readers
to read from the pipe to the stdout/stderr files, link the command
output directly to the files. This simplifies things and avoids some
race conditions where the command finishes but the readers haven't
copied everything to the pipes yet.

A new method ModuleMgr::FailResult is used to send a failed result. It
creates a result with successful=false and failure_details filled in
with the full error string.

In the scrapers, set Successful=true.

* Unit test for unsuccessful comp results

Test changes for failed compliance results:

1. Modify test-module to optionally exit with a non-zero value, which
will result in a failed task.

2. Modify the test to support test definitions with unsuccessful
results. In those cases, a regex is matched against the failure details.

3. Add 2 new test cases:

 - exit_failure looks for a test module with non-zero exit code.
 - fail_module looks for running a module that is not executable and
   will not run.

There are other error paths that we may want to write unit tests for,
but some may be hard to trigger (not being able to create pipes, etc.)

Also improve the tests generally a bit:

- Also verify pass count in addition to tests run count.
- Verify machine-id/customer-id are set in results

* Update to kube-bench 0.0.17.1

This includes a fix to improve master node detection.