Add option to disable -Werror for bootstrap-agent (#1110)
SMAGENT-1175 introduced a change that enables both -Wextra and -Werror.
This works fine for container builds because we have suppressions for
any warnings that are generated by the compiler in that environment.
There are, however, use cases for building the agent directly on the
host.  The host may have a newer compiler with a different set of
warnings, and -Werror will cause those to trigger build failures.
This change does two thing: (1) Adds an option BUILD_WARNINGS_AS_ERRORS
to the agent CMakeLists.txt that controls whether or not -Werror is
enabled, and (2) Updates bootstrap-agent to conditionally enable the
option based on an environment variable with the same name.  If the
environment variable exists, bootstrap-agent will use that value;
otherwise it will set it to ON.

Anyone who wants to run bootstrap-agent outside of the container build
environment will want to set BUILD_WARNINGS_AS_ERRORS=OFF.