Memdump test improvements (#1144)
* Improve error reporting

In several places related to capture job handling/memdump
initialization, errors didn't have enough context or tests didn't
provide details on test failures. Add necessary info.

(This won't exactly compile as-is, as the changes are intertwined with a
later commit).

* Retry memdump init posix_fallocate() on EINTR

One thing the memdumper does when it initializes is to preallocate the
shared memory region using `posix_fallocate()`. This can be interrupted
by a signal, causing fallocate to return EINTR.  In our case, it was
gperftools' profiler, which sends repeated SIGPROF signals.

To help mitigate this, add a configurable number of retries to
posix_fallocate() before giving up. This is not foolproof--if profiling
is enabled and SIGPROF signals are delivered faster than
posix_fallocate() will complete, the init will always fail.

* Make sure analyzer cpu profiling is off by default

In some unit tests, cpu profiling was being enabled unintentially, and
not being disabled before running a following test.

The reason profiling was being enabled is that the analyzer's bool
`m_dragent_cpu_profile_enabled` wasn't explicitly being set to
false. Normally, this gets handled when the dragent's configuration is
copied to the analyzer's configuration--the dragent configuration class
does set `m_dragent_cpu_profile_enabled` to false in
`dragent_configuration::init`.

However, lots of unit tests create an analyzer without any corresponding
dragent configuration, the bool is not initialized, and might be set to
true.

So explicitly set it to false, and also when profiling is enabled,
always stop profiling in the analyzer's destructor.

* Remove unused code

Lll of the saturation inactivity stuff was left over from a previous
version of the memdumper that allowed writing to multiple files at
once. Now it's handled in the capture job handler. Remove it.