Support new falco rules format in policies message
When receiving a policies message, look at the falco_group field to find
the default and custom rules sections. If present, read the list of
files and variants, choosing the variant that has the largest required
engine version that is compatible with our falco engine.

If no default rules/custom rules section is present, fall back to the
(single) system/user rules file from the falco_rules.content list.