Audit tap support. SMAGENT-1232
* Record all connection state changes

* Collect all connection state and exit events in audit tap

* Report audit data to collector

* Add config option to enable audit tap

* Try to send environments in future audit tap flushes

When we go over the limit of environments per flush, remember
the pids that we failed to send and try to send them later.

We need to handle this manually (unlike the env dump in the normal
flush process), because once we get a connection event, we may never
get another one for the same process.

Old environments are sent only after new ones, while the original
process is still alive.

* Debug only mode for audit tap

* Only send audit tap data after we send some metrics

The collector authenticates the agent by the first metrics message,
so it has to be sent before any audit tap messages.

* Send timestamp, machine_id and hostname in audit tap messages. SMAGENT-1232