Ensure we have enough POSIX queue space. SMAGENT-1104 (#1118)
* Switch to private IPC namespace by default. SMAGENT-1104

We don't switch the namespace when running inside a container
(we already have a private IPC namespace there) and when
reexecuting the agent binary.

* Set RLIMIT_MSGQUEUE only once, in the main process. SMAGENT-1104

The subprocesses (sdchecks, sdjagent etc.) will then inherit the limit.

* Make the RLIMIT_MSGQUEUE value configurable. SMAGENT-1104

We don't allow the value to go below the minimum (the previously
hardcoded value). Unfortunately, logging isn't enabled while
parsing the config so we can't easily log a warning message
without e.g. modifying the value in print_configuration(),
which seems a bit too surprising.