Adding TLS support for K8s Audit Server
Through dragent.yaml config file, the endpoint can be configured to be
either HTTP or HTTPS. At this point, only one of the two can be
chosen. If the k8s_audit_server_tls_enabled is set to true, the
filename of a X509 public key and certificate needs to be provided.

The configuration has been manually tested, as currently I haven't yet
found test suits that check for dragent configuration. Just for
reference sake, here's a sample dragent.yaml (with current default
values)

security:
	k8s_audit_server_enabled: true
	k8s_audit_server_url: "localhost"
	k8s_audit_server_port: 8765
	k8s_audit_server_tls_enabled: false
	k8s_audit_server_x509_cert_file: ""
	k8s_audit_server_x509_key_file:  ""