SMAGENT-1249: Enable host statsd server (#1166)
* SMAGENT-1249: Enable host statsd server

This change introduces a new configuration option:
'statsd.use_host_statsd'.  With this config set, the agent will start
the statsite server with the UDP and TCP ports configured as 0.  With
that configuration, statsite will not listen for incoming TCP or UDP
connections on any port.  Additionally, this change updates the "statsd
sniffing" logic to process data sent to localhost if the use_host_statsd
option is set.

With this configuration, customers can run their own statsd server
listening on port 8125.  We'll sniff that traffic, send it to our
statsite server, and process the events in the same way that we do with
the other sniffing paths.

* Address Greg's comments:
    * Fix a incorrect comment about the result of using 0 as the port
    * Remove testing entry from dragent.default.yaml.