Set CRI socket path unconditionally. SMAGENT-1256
CRI socket path defaults to /run/containerd/containerd.sock in sysdig
but the agent does its own detection and when the socket is not found,
set it to an empty string.

This is needed so that we don't keep trying to connect to containerd.sock
for containerd events if the socket doesn't exist.