Add ability to save temp files from comp tasks (#1197)
A new config option compliance: {compliance_save_temp_files: <bool>}
controls whether the temporary directory used for a compliance task's
output is removed right away. The default is false, so the directory is
removed immediately after the task runs. When set to true, the temporary
path is save and is either removed upon Stop() or when the next task
runs.

This fixes SSPROD-1130.