Run tasks now (#1176)
* Add ability to run comp tasks on demand

Define a new message COMP_RUN (21) that can be sent from the collector
and includes a list of task ids. This corresponds to a new cointerface
RPC RunTasks. RunTasks just creates a ScheduledTask for each provided
task id and calls its RunNow method. This does block the RPC from the
main dragent but that RPC is done asynchronously so it doesn't block the
main agent process.

The results are sent on the channels associated with the module manager
and the Start() RPC so they are returned like any other result.

* Unit test for run now

This test starts a task based on a calendar that would run once every
hour. It waits for the first result, clears the set of results, and then
sends a run_tasks message to run the first task immediately.

* Refresh comp tasks right after getting a calendar

Previously, the security manager would periodically check for an updated
calendar message (120 secs). If the calendar had changed, the calendar
would be loaded.

Eliminate this 120 second delay by setting a flag in
security_mgr::set_compliance_calendar() and checking the flag in
security_mgr's periodic tasks callback. This allows the tasks to be
loaded right after being received.

* Make sure only 1 module can run at a time

With the introduction of run now, there's a chance that someone could
request that a scheduled task runs at the same time that someone ran a
task manually.

Serialize this through a goroutine RunModules(). A channel associated
with each module receives requests to run a module. The return
status (whether the module was started successfully) is returned on a
response channel.

The context is passed down through the goroutines so everything is
cancelled on Stop().

* Reeval compliance tasks on orchestrator evts msg

When the agent receives an ORCHESTRATOR_EVENTS message, this may result
in an updated set of hosts (including the local host). When this occurs,
also reevaluate the set of compliance tasks to run, using a new method
request_refresh_compliance_tasks(). This sets
m_should_refresh_compliance_tasks (formerly m_new_compliance_calendar),
which is the trigger in the security manager's periodic tasks job to
call refresh_compliance_tasks().

This is part of the fix for SSPROD-1407.

* Reeval registered scopes on hosts metadata changes

In refresh_hosts_metadata(), reevaluate any registered scopes that have
a hostname component and reevaluate them.

This is part of the fix for SSPROD-1407.

* Track tasks to run by id, not name

That way they are unaffected by task name changes.

This fixes SSPROD-1406.