[SMAGENT-1443] Create sinsp_mock and use it to generate fake events in unit test (#1290)
Create sinsp_mock and the event builder and use them to generate fake events in unit test. This should be the model to test a large part of libsanalyzer.

This was done by creating the dragent.test and libsanalyzer.test libraries where code can be injected with the SYSDIG_TEST define.