Smagent 1416 cointerface log msgs (#1298)
* Handle subprocess log lines > 4k

Handle really long log lines from subprocesses by keeping a buffer of
pending log content. Reads are still done in up to 4k chunks, but after
reading each chunk it looks for a newline. Once a newline is found, the
complete line is passed to the designated parsing function.

To support this pending state, the pair<parsing func,watchdog state> was
changed to a state struct, adding the pending log buffer.

This fixes SMAGENT-1469 and SMAGENT-1416.

* Log this at debug level

It pollutes the logs otherwise.

* Minor fix: transposed two members in an inizializer list