[SMAGENT-1460] small reorg for Nathan's fake-collector (#1316)
* small reorg for Nathan's fake-collector

* Modified UT -> fake_collector protocol

Travis was encountering the problem that a read() on the fake collector
would receive multiple messages from the connection manager (due to
general Travis slowness).

I addressed this problem by prepending the length of the string to the
start of the message and then reading only that length from the socket
and no more. This appears to have solved the problem.