SMAGENT-1466: Unfriendly sinsp_analyzer (#1311)
* SMAGENT-1466: Unfriendly sinsp_analyzer

This change eliminates all friends from `sinsp_analyzer`.  Much of this
change involves adding helper methods to `sinsp_analyzer` to expose some
information from some of its internal details.

There are a couple of cases where I've made previously private methods
public.  This is less than ideal, but for now I think it's better than
having the friends.  Hopefully as we continue to refactor this, we can
reorganize things so that that's no longer needed.

* Update baseliner integration test

The changes I made to the falso baseliner needed a corresponding
client-side change in the baseliner integration test.

* Address comments from walk-through

* Resolve some issues introduced by the merge