Async container info [SMAGENT-1324] (#1251)
* Update for new use of notify_new_container.

* Look for container event for container info

Now that container info is only guaranteed to be available at a
CONTATINER_JSON event, look for that event in the docker container test.

* Destroy sinsp before recreating

This ensures that static objects are deleted before being recreated.

* Update build to include tbb when needed

Also use a consistent TBB_INCLUDE_DIR so it can be used interchangably
between sysdig, falco, and the agent.

* Allow unknown containers when a cri lookup fails

This can occur now that docker container lookups occur in the background.

* Use CONTAINER_JSON events for container policies.

* Get unit tests passing

The big change was that for activity that occurs in containers, any
scope-based policy using the container name/image/etc won't work until
the full container metadata has been fetched. To address that, change
the activity so it occurs in a loop so it occurs multiple times. When
necessary, move away from touch <file> towards echo > <file> so the file
is actually opened each time.

This also slightly changed the baselines to allow for additional
programs like sleep, echo, etc.

Also ensure that containers for the container prefix tests are slightly
longer lived so they exist long enough for the metadata lookup and the
container event to be emitted.

A smaller change was to not consider a "container" event a syscall, so
you don't need to add it to baselines, etc.

Also a couple of small other test improvements:

 - In some of the filesystem tests, actually make sure that the program
   "tests" is doing the writes and not the "echo" inside a container.
 - If the docker swarm test can't read any policy events, don't try to
   dereference the pointer.

* Clear the policy cache when a container is added

It can now be the case that a container is added to infrastructure state
multiple times--once with a conainer with incomplete metadata, and again
later with complete metadata. This causes problems with the scope
cache, where scope lookups for the earlier container with incomplete
metdata are used even though the container now has additional metadata
that might result in a different scope match.

To fix this, make sure to clear the policy cache for the container id
every time a container is added.

* Missed a couple of short-lived containers

They, like the others, need to run for a bit so the threadinfo exists
when the container exists.

* Fix indenting to only use spaces.

* Get rid of CT_UNKNOWN

m_metadata_complete is now the way to identify that container info is incomplete.

* Further document why clear_cached_result is needed

It's because match_scope always checks the cache first.

* Add tbb dependency

Now required from libsinsp.

* Add tbb as a dependency

Required by libsinsp now.