SMAGENT-1527: Protobuf Dump Broken (#1339)
* SMAGENT-1527: Protobuf Dump Broken

In SMAGENT-1396 I introduced a regression that broke protobuf dumping.
The cause was that didn't account for the fact that the
`sinsp_configuration` object that the `sinsp_analyzer` has a pointer to
would change outside of the flow control in the analyzer.  I organized
the code so that when the analyzer's config update method was invoked,
that it passed the config of interest to the protobuf serializer.
During execution, after the call to the update API, the original config
object gets updated, and the protobuf serializer does not receive those
updates.

This change updates the `metric_serializer` (and, in turn, the
`protobuf_metric_serializer`) to maintain a pointer to the analyzer's
configuration.  This way, when the analyzer's configuration changes,
the serializer will get those changes.

* space vs tabs