Fix analyzer config management (#1346)
* Fix analyzer config management

An earlier change that I introduced broke configuration
for the `sinsp_analyzer`.  This chagne restores what was there before.
The original change was intended to avoid a potential race, but on more
careful consideration, I don't think that race is possible.  The method
to update the configuraiton will throw an exception if a capture is in
progress.