Update docker bench (#1341)
* Update docker-bench-security to 1.3.4+item limits

Update docker-bench-security to 1.3.4, plus the changes in
https://github.com/draios/docker-bench-security/pull/2.

* Limit the number of reported items to 500

In some evironments, there may be a very large number of images,
containers, etc not satisfying a given test. For example, in one
environment, we saw **378k** images not satisfying 4.6, mostly because
the customer was never cleaning up old images.

To avoid excess lists of items, use the new config item introduced in
https://github.com/draios/docker-bench-security/pull/2 to limit the
number of reported items to 500.

This fixes https://sysdig.atlassian.net/browse/SSPROD-1562.