SMAGENT 1506: Remove unnecessary serialization (#1334)
* SMAGENT 1506: Remove unnecessary serialization

When the emit_metrics_to_file config option is set, we write protobufs
to file.  It looks like once upon a time, we manually serialized the
protobuf to a byte array, either compressed or not, then wrote that to
a file.  Currently, we just write the output of the `DebugString` method.
However, we still go through the process of serializing (and maybe
compressing) the protobuf to a byte array (but we don't use the result).
This change does the following:

* Removes the unneeded manual serialization to the byte array
* Removes the 'compress' config option for the analyzer.  Note that
  the "real" serialization process checks the config option in dragent's
  config, and that I did not touch that.
* Removes the state and APIs for tracking whether compressed buffers are
  used in protobuf_metric_serializer.
* Touched up all the files referencing those APIs.