SMAGENT-1562: Make UTs valgrind clean again (#1359)
When we initially introduced Valgrind into the build, all UTs built
clean (for both Valgrind and Helgrind). Since then a number of
additional UTs have been added that are not clean. That has grown
to the point that it causes unexpected failures in Travis, slowing
down development.

This change makes all the UTs run cleanly again.  Some where fixed.
The following were disabled:

* sinsp_worker_test.end_to_end_basic
* connection_manager_test.failure_to_connect
* connection_manager_test.connect_transmit
* analyzer_test.end_to_end_basic
* analyzer_test.coalesce_containers_test