SMAGENT-1573: go_expvar not returning metrics (#1367)
As the `go_expvar` app check interprets the data that it receives from
the application that it's monitoring, it attempts to determine if a
value can be converted to a floating-point number with:

```python
try:
    float(value)
except ValueError:
    self.log.debug("Unreportable value for path %s: %s" % (path, value))
    continue
```

When `value` isn't a float, a `TypeError` exception is thrown, not a
`ValueError`.  This change updates that check to handle both `TypeError`
and `ValueError`, and decreases the severity of the associated log
message.