SMAGENT-1559: Separate logging concerns (#1358)
* SMAGENT-1559: Separate logging concerns

Currently `sinsp_logger` is responsible for handling various unrelated
tasks:

* trace logging
* user event logging
* memdump logging

That then translates to `dragent_logger` being responsible for handling
the same events.  This change decouples the concerns.  Why? When
modularizing the `sinsp_analyzer`, I'd like to isolate the user event
handling, and I don't want that to be coupled with the trace logger.
This change will also enable us to move the `dragent_logger` from the
`dragent` component to the `userspace-shared` component in the future
(which is something we'd like to eventually do).