[SMAGENT-1590] Change connect workflow
There was a potential race condition where the connect thread could be
left holding a dangling reference.

I'm tired of fixing these problems so I went ahead and changed the
workflow to guarantee the connect thread finishes before the function
does, at the risk of getting killed by the watchdog in weird
circumstances.