SMAGENT-1473 send raw prometheus metrics (#1389)
Add support to send the raw values we scrape for prometheus metrics to
the backend, in order to support PromQL.  Related to protobuf and
backend work for raw collection in SMBACK-2855 and SMBACK-2856.

Adds two new prometheus configuration parameters, ingest_raw and
ingest_calculated. ingest_calculated is what we have been doing up until
now; ingest_raw toggles the new behavior. Both may be set to true, in
which case, we'll report the existing calculated metrics, as well as the
raw metrics.

The raw metrics are reported with the name seen during scraping, and
identified with a new app metric type. Note that we will collect
similarly named metrics if both ingest_raw and ingest_calculated are
enabled. For example: if a prometheus endpoint has a counter called
"sample_counter", and both flags are true, the protobuf will contain a
metric called sample_counter, whose type is an app metric gauge, whose
value is the agent-caculated rate; the protobuf will also contain an app
metric called sample_counter, whose type is the new prometheus_raw app
metric type, and whose value is the last value we collected during
scraping.