SMAGENT-1698: Kernel support for statsd port conf (#1453)
* SMAGENT-1698: Kernel support for statsd port confi

This change refactors the Agent's management of statsd configuration
into a new namespace: `statsite_config`.  When the configured statsd udp
port is not the standard port 8125, then it calls into sysdig to push
the newly-configured port to the kernel (either the driver or the eBPF
probe).  This change also increases the maximum limit for statsd metrics
to 3000.