SMAGENT-1497: Pull emit_statsd() out of analyzer (#1457)
* wip

* SMAGENT-1497: Pull emit_statsd() out of analyzer

This change pulls statsd emission out of `sinsp_analyzer` and into a
separate class (statsite_statsd_emitter).  That implementation is a
realization of a new interface, `statsd_emitter`.  There's a
null-object-based implementation of the interface `null_statsd_emitter`
that is used when statsd is disabled.

This change adds interface `statsd_stats_source` to the statsite_proxy
so that we can put an abstraction over what object is actually producing
the stats.

I added a `dummy_statsd_stats_source` test realization of the
`statsd_stats_source` so that we can generate fake statsd metrics.  I
use that dummy implementation to test `statsite_statsd_emitter`.

This change also add a new UT helper, `scoped_file_pointer` as an RAII
wrapper over `FILE*`.

Finally, this change updates the `main()` of the libsanalyzer unit test
to clean up protobuf-based memory before terminating the test.